Changes for connection retry in case of db server restart fo prompt service
